requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
sentence-transformers>=2.2.2
torch>=2.0.0
numpy>=1.24.3
Werkzeug>=3.0.1
Pillow>=10.0.0
python-dotenv>=1.0.0
